+++ mlflow/deployments/server/providers/bedrock.py
@@ -13,10 +13,10 @@ from mlflow.gateway.constants import (
     MLFLOW_AI_GATEWAY_ANTHROPIC_DEFAULT_MAX_TOKENS,
 )
 from mlflow.gateway.exceptions import AIGatewayConfigException
-from mlflow.gateway.providers.anthropic import AnthropicAdapter
-from mlflow.gateway.providers.base import BaseProvider, ProviderAdapter
-from mlflow.gateway.providers.cohere import CohereAdapter
-from mlflow.gateway.providers.utils import rename_payload_keys
+from mlflow.deployments.server.providers.anthropic import AnthropicAdapter
+from mlflow.deployments.server.providers.base import BaseProvider, ProviderAdapter
+from mlflow.deployments.server.providers.cohere import CohereAdapter
+from mlflow.deployments.server.providers.utils import rename_payload_keys
 from mlflow.gateway.schemas import completions
 
 AWS_BEDROCK_ANTHROPIC_MAXIMUM_MAX_TOKENS = 8191
