+++ mlflow/deployments/server/providers/cohere.py
@@ -6,8 +6,8 @@ from fastapi import HTTPException
 from fastapi.encoders import jsonable_encoder
 
 from mlflow.gateway.config import CohereConfig, RouteConfig
-from mlflow.gateway.providers.base import BaseProvider, ProviderAdapter
-from mlflow.gateway.providers.utils import rename_payload_keys, send_request, send_stream_request
+from mlflow.deployments.server.providers.base import BaseProvider, ProviderAdapter
+from mlflow.deployments.server.providers.utils import rename_payload_keys, send_request, send_stream_request
 from mlflow.gateway.schemas import chat, completions, embeddings
