+++ mlflow/deployments/server/providers/huggingface.py
@@ -5,8 +5,8 @@ from fastapi import HTTPException
 from fastapi.encoders import jsonable_encoder
 
 from mlflow.gateway.config import HuggingFaceTextGenerationInferenceConfig, RouteConfig
-from mlflow.gateway.providers.base import BaseProvider
-from mlflow.gateway.providers.utils import (
+from mlflow.deployments.server.providers.base import BaseProvider
+from mlflow.deployments.server.providers.utils import (
     rename_payload_keys,
     send_request,
 )
