+++ mlflow/deployments/server/providers/mistral.py
@@ -4,8 +4,8 @@ from typing import Any, Dict
 from fastapi.encoders import jsonable_encoder
 
 from mlflow.gateway.config import MistralConfig, RouteConfig
-from mlflow.gateway.providers.base import BaseProvider, ProviderAdapter
-from mlflow.gateway.providers.utils import send_request
+from mlflow.deployments.server.providers.base import BaseProvider, ProviderAdapter
+from mlflow.deployments.server.providers.utils import send_request
 from mlflow.gateway.schemas import completions, embeddings
