+++ mlflow/deployments/server/providers/mosaicml.py
@@ -7,8 +7,8 @@ from fastapi.encoders import jsonable_encoder
 
 from mlflow.exceptions import MlflowException
 from mlflow.gateway.config import MosaicMLConfig, RouteConfig
-from mlflow.gateway.providers.base import BaseProvider
-from mlflow.gateway.providers.utils import rename_payload_keys, send_request
+from mlflow.deployments.server.providers.base import BaseProvider
+from mlflow.deployments.server.providers.utils import rename_payload_keys, send_request
 from mlflow.gateway.schemas import chat, completions, embeddings
