+++ tests/gateway/providers/test_ai21labs.py
@@ -8,7 +8,7 @@ from fastapi.encoders import jsonable_encoder
 from mlflow.exceptions import MlflowException
 from mlflow.gateway.config import RouteConfig
 from mlflow.gateway.constants import MLFLOW_GATEWAY_ROUTE_TIMEOUT_SECONDS
-from mlflow.gateway.providers.ai21labs import AI21LabsProvider
+from mlflow.deployments.server.providers.ai21labs import AI21LabsProvider
 from mlflow.gateway.schemas import chat, completions, embeddings
 
 from tests.gateway.tools import MockAsyncResponse
