+++ tests/gateway/providers/test_bedrock.py
@@ -10,7 +10,7 @@ from mlflow.gateway.config import (
     AWSRole,
     RouteConfig,
 )
-from mlflow.gateway.providers.bedrock import AmazonBedrockModelProvider, AmazonBedrockProvider
+from mlflow.deployments.server.providers.bedrock import AmazonBedrockModelProvider, AmazonBedrockProvider
 from mlflow.gateway.schemas import completions
 
 from tests.gateway.providers.test_anthropic import (
@@ -383,7 +383,7 @@ async def test_bedrock_request_response(
     provider, config, payload, response, expected, model_request, aws_config
 ):
     with mock.patch("time.time", return_value=1677858242), mock.patch(
-        "mlflow.gateway.providers.bedrock.AmazonBedrockProvider._request", return_value=response
+        "mlflow.deployments.server.providers.bedrock.AmazonBedrockProvider._request", return_value=response
     ) as mock_request:
         if not expected:
             pytest.skip("no expected value")
