+++ tests/gateway/providers/test_cohere.py
@@ -8,7 +8,7 @@ from pydantic import ValidationError
 
 from mlflow.gateway.config import RouteConfig
 from mlflow.gateway.constants import MLFLOW_GATEWAY_ROUTE_TIMEOUT_SECONDS
-from mlflow.gateway.providers.cohere import CohereProvider
+from mlflow.deployments.server.providers.cohere import CohereProvider
 from mlflow.gateway.schemas import chat, completions, embeddings
 
 from tests.gateway.tools import MockAsyncResponse, MockAsyncStreamingResponse
