+++ tests/gateway/providers/test_mistral.py
@@ -9,7 +9,7 @@ from pydantic import ValidationError
 
 from mlflow.gateway.config import RouteConfig
 from mlflow.gateway.constants import MLFLOW_GATEWAY_ROUTE_TIMEOUT_SECONDS
-from mlflow.gateway.providers.mistral import MistralProvider
+from mlflow.deployments.server.providers.mistral import MistralProvider
 from mlflow.gateway.schemas import completions, embeddings
 
 from tests.gateway.tools import MockAsyncResponse
