+++ tests/gateway/providers/test_openai.py
@@ -9,7 +9,7 @@ from pydantic import ValidationError
 from mlflow.exceptions import MlflowException
 from mlflow.gateway.config import OpenAIConfig, RouteConfig
 from mlflow.gateway.constants import MLFLOW_GATEWAY_ROUTE_TIMEOUT_SECONDS
-from mlflow.gateway.providers.openai import OpenAIProvider
+from mlflow.deployments.server.providers.openai import OpenAIProvider
 from mlflow.gateway.schemas import chat, completions, embeddings
 
 from tests.gateway.tools import (
