+++ tests/gateway/providers/test_palm.py
@@ -8,7 +8,7 @@ from pydantic import ValidationError
 
 from mlflow.gateway.config import RouteConfig
 from mlflow.gateway.constants import MLFLOW_GATEWAY_ROUTE_TIMEOUT_SECONDS
-from mlflow.gateway.providers.palm import PaLMProvider
+from mlflow.deployments.server.providers.palm import PaLMProvider
 from mlflow.gateway.schemas import chat, completions, embeddings
 
 from tests.gateway.tools import MockAsyncResponse
