+++ tests/gateway/providers/test_provider_utils.py
@@ -1,6 +1,6 @@
 import pytest
 
-from mlflow.gateway.providers.utils import (
+from mlflow.deployments.server.providers.utils import (
     rename_payload_keys,
 )
