+++ tests/gateway/test_integration.py
@@ -9,17 +9,17 @@ import mlflow.gateway.utils
 from mlflow.exceptions import MlflowException
 from mlflow.gateway import MlflowGatewayClient, get_route, query, set_gateway_uri
 from mlflow.gateway.config import Route
-from mlflow.gateway.providers.ai21labs import AI21LabsProvider
-from mlflow.gateway.providers.anthropic import AnthropicProvider
-from mlflow.gateway.providers.bedrock import AmazonBedrockProvider
-from mlflow.gateway.providers.cohere import CohereProvider
-from mlflow.gateway.providers.huggingface import HFTextGenerationInferenceServerProvider
-from mlflow.gateway.providers.mistral import MistralProvider
-from mlflow.gateway.providers.mlflow import MlflowModelServingProvider
-from mlflow.gateway.providers.mosaicml import MosaicMLProvider
-from mlflow.gateway.providers.openai import OpenAIProvider
-from mlflow.gateway.providers.palm import PaLMProvider
-from mlflow.gateway.providers.togetherai import TogetherAIProvider
+from mlflow.deployments.server.providers.ai21labs import AI21LabsProvider
+from mlflow.deployments.server.providers.anthropic import AnthropicProvider
+from mlflow.deployments.server.providers.bedrock import AmazonBedrockProvider
+from mlflow.deployments.server.providers.cohere import CohereProvider
+from mlflow.deployments.server.providers.huggingface import HFTextGenerationInferenceServerProvider
+from mlflow.deployments.server.providers.mistral import MistralProvider
+from mlflow.deployments.server.providers.mlflow import MlflowModelServingProvider
+from mlflow.deployments.server.providers.mosaicml import MosaicMLProvider
+from mlflow.deployments.server.providers.openai import OpenAIProvider
+from mlflow.deployments.server.providers.palm import PaLMProvider
+from mlflow.deployments.server.providers.togetherai import TogetherAIProvider
 from mlflow.utils.request_utils import _cached_get_request_session
 
 from tests.gateway.tools import (
